refactor: split into separate script
