+++ .github/workflows/pull-project-docs.yml
@@ -33,16 +33,20 @@ jobs:
           echo "project_repository=${{ github.event.client_payload.project_repository || inputs.project_repository }}" >> $GITHUB_OUTPUT
 
       - name: Pull project documentation
-        id: pull-docs
         run: |
-          deno task pull-project-docs --project "${{ steps.params.outputs.project_repository }}" --token "${{ steps.generate_token.outputs.token }}"
+          deno task pull-project-docs --project "${{ steps.params.outputs.project_repository }}"
+
+      - name: Fetch project commit
+        id: fetch-commit
+        run: |
+          deno task fetch-project-commit --project "${{ steps.params.outputs.project_repository }}" --token "${{ steps.generate_token.outputs.token }}"
 
       - name: Create pull request
         uses: dfinity/ci-tools/actions/create-pr@main
         with:
           branch_name: "build/docs-update"
           base_branch_name: "main"
-          pull_request_title: "build: update project docs from ${{ steps.params.outputs.project_repository }}@${{ steps.pull-docs.outputs.commit_sha }}"
-          pull_request_body: "This pull request was automatically created to update the project documentation from ${{ steps.params.outputs.project_repository }}@${{ steps.pull-docs.outputs.commit_sha }}."
-          commit_message: "build: update docs from ${{ steps.params.outputs.project_repository }}@${{ steps.pull-docs.outputs.commit_sha }}"
+          pull_request_title: "build: update project docs from ${{ steps.params.outputs.project_repository }}@${{ steps.fetch-commit.outputs.commit_sha }}"
+          pull_request_body: "This pull request was automatically created to update the project documentation from ${{ steps.params.outputs.project_repository }}@${{ steps.fetch-commit.outputs.commit_sha }}."
+          commit_message: "build: update docs from ${{ steps.params.outputs.project_repository }}@${{ steps.fetch-commit.outputs.commit_sha }}"
           token: ${{ steps.generate_token.outputs.token }}
